+++ bin/faucet/src/server/mod.rs
@@ -1,9 +1,6 @@
 use std::{
     collections::HashSet,
-    sync::{
-        Arc,
-        atomic::{AtomicU64, Ordering},
-    },
+    sync::{Arc, atomic::AtomicU64},
     time::{SystemTime, UNIX_EPOCH},
 };
 
@@ -29,7 +26,7 @@ use url::Url;
 use crate::{
     COMPONENT,
     faucet::FaucetId,
-    server::{get_note::get_note, get_pow::get_pow, get_tokens::MintRequestError},
+    server::{get_note::get_note, get_pow::get_pow},
     types::AssetOptions,
 };
 
@@ -41,7 +38,7 @@ mod get_pow;
 mod get_tokens;
 mod pow;
 pub use api_key::ApiKey;
-pub use get_tokens::{MintRequest, MintResponse, MintResponseSender};
+pub use get_tokens::{MintRequest, MintRequestError, MintResponse, MintResponseSender};
 pub use pow::PoWConfig;
 
 // FAUCET STATE
@@ -64,7 +61,7 @@ impl Server {
     pub fn new(
         faucet_id: FaucetId,
         max_supply: u64,
-        claimed_supply: u64,
+        claimed_supply: Arc<AtomicU64>,
         asset_options: AssetOptions,
         mint_request_sender: MintRequestSender,
         pow_secret: &str,
@@ -76,7 +73,7 @@ impl Server {
         let metadata = Metadata {
             id: faucet_id,
             asset_amount_options: asset_options,
-            claimed_supply: Arc::new(AtomicU64::new(claimed_supply)),
+            claimed_supply,
             max_supply,
         };
         // SAFETY: Leaking is okay because we want it to live as long as the application.
@@ -184,11 +181,6 @@ impl Server {
             .expect("current timestamp should be greater than unix epoch")
             .as_secs();
         self.pow.submit_challenge(account_id, api_key, challenge, nonce, timestamp)
-    }
-
-    /// Increments the claimed supply counter by the given amount.
-    pub(crate) fn increment_claimed_supply(&self, amount: u64) {
-        self.metadata.claimed_supply.fetch_add(amount, Ordering::Relaxed);
     }
 }