+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Redesigned the home frontend ([#20](https://github.com/0xMiden/miden-faucet/pull/20)).
 - Redesigned the tokens request flows ([#25](https://github.com/0xMiden/miden-faucet/pull/25)).
 - Added faucet supply amounts to the metadata ([#30](https://github.com/0xMiden/miden-faucet/pull/30)).
+- Added supply exceeded check ([#31](https://github.com/0xMiden/miden-faucet/pull/31)). 
 
 ## 0.10.0 (2025-07-10)